build(workspace): migrate to Rust edition 2024
Workspace configuration
- Set edition = "2024", rust-version = "1.96", and resolver = "3" in the root [workspace.package]/[workspace], inherited by members via field.workspace = true.
- Centralize internal path-crate and external dependency versions under [workspace.dependencies], keeping a default-features = false baseline that members opt back into locally.
- Declare the recognized custom cfgs (tracing_unstable, flaky_tests, unsound_local_offset) under [workspace.lints.rust] and inherit lints across members.

Dependencies and edition adaptations
- Upgrade all dependencies to their latest compatible versions (cargo upgrade --recursive --incompatible --pinned).
- Apply edition-2024 rustfmt formatting and clippy auto-fixes across crates, examples, tests, and benches.
- Adapt source to upgraded crate APIs and edition-2024 idioms (tracing macros and level filters, tracing-mock subscriber, tracing-journald, tracing-log, tracing-subscriber filter/fmt).

Feature and structural fixes
- Implement the documented tracing-futures tokio-executor feature: split executor/futures_01.rs into a tokio_executor module (Executor/TypedExecutor impls via the standalone crate) and a tokio_runtime module (Runtime/current_thread conveniences), and make tokio a strict superset that implies tokio-executor.
- Normalize the tracing-tower feature table to explicit dep:-based features (tower-layer, tower-make, http).
- Rename the non-workspace test package to test_static_max_level_features to match its directory, and add tokio-executor to the tracing-futures minimal-versions CI exclude.

Documentation and project memory
- Add a root AGENTS.md plus per-crate AGENTS.md files, each paired with a CLAUDE.md stub importing it.
- Re-align fork identity to the strict-rs org (repository fields, issue_tracker_base_url, CI badge URLs) while preserving tokio-rs upstream attribution and assets.
- Raise the documented MSRV to 1.96 across READMEs and crate rustdoc, and dedupe the tracing README supported-versions section.

+++ tracing-futures/src/stdlib.rs
@@ -22,8 +22,8 @@ mod no_std {
 
     pub(crate) use core::{
         any, array, ascii, cell, char, clone, cmp, convert, default, f32, f64, ffi, future, hash,
-        hint, i128, i16, i8, isize, iter, marker, mem, num, ops, option, pin, ptr, result, task,
-        time, u128, u16, u32, u8, usize,
+        hint, i8, i16, i128, isize, iter, marker, mem, num, ops, option, pin, ptr, result, task,
+        time, u8, u16, u32, u128, usize,
     };
 
     pub(crate) mod borrow {
